+++ src/datavisualization/utils/utils_p.h
@@ -67,7 +67,7 @@ public:
     static QImage getGradientImage(const QLinearGradient &gradient);
 
     static ParamType findFormatParamType(const QString &format);
-    static QString formatLabel(const QByteArray &format, ParamType paramType, float value);
+    static QString formatLabel(const QByteArray &format, ParamType paramType, qreal value);
     static QString defaultLabelFormat();
 
     static float wrapValue(float value, float min, float max);
